+++ src/main/java/com/tavemakers/surf/domain/member/controller/AdminMemberController.java
@@ -38,7 +38,7 @@ public class AdminMemberController {
     @Operation(summary = "회원 역할 변경", description = "여러 회원들의 역할을 변경합니다.")
     @PatchMapping("/v1/admin/members/role")
     public ApiResponse<Void> changeMembersRole(
-            @RequestBody RoleChangeReqDTOV2 dto
+            @RequestBody @Valid RoleChangeReqDTOV2 dto
     ) {
         memberAdminUsecase.changeMembersRole(dto);
         return ApiResponse.response(HttpStatus.OK, "회원 역할이 성공적으로 변경되었습니다.",null);
