+++ src/main/java/com/tavemakers/surf/domain/notification/repository/DeviceTokenRepository.java
@@ -3,6 +3,9 @@ package com.tavemakers.surf.domain.notification.repository;
 import com.tavemakers.surf.domain.notification.entity.DeviceToken;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
@@ -12,6 +15,8 @@ public interface DeviceTokenRepository extends JpaRepository<DeviceToken, Long> 
 
     List<DeviceToken> findAllByMemberIdAndEnabledTrue(Long memberId);
 
-    void deleteByMemberId(Long memberId);
+    @Modifying(clearAutomatically = true, flushAutomatically = true)
+    @Query("DELETE FROM DeviceToken d WHERE d.memberId = :memberId")
+    void deleteByMemberId(@Param("memberId") Long memberId);
 
 }
