+++ src/main/java/com/tavemakers/surf/domain/notification/repository/NotificationRepository.java
@@ -30,5 +30,7 @@ public interface NotificationRepository extends JpaRepository<Notification, Long
 
     boolean existsByIdAndMemberId(Long id, Long memberId);
 
-    void deleteByMemberId(Long memberId);
+    @Modifying(clearAutomatically = true, flushAutomatically = true)
+    @Query("DELETE FROM Notification n WHERE n.memberId = :memberId")
+    void deleteByMemberId(@Param("memberId") Long memberId);
 }
