+++ src/main/java/com/tavemakers/surf/domain/reservation/repository/ReservationRepository.java
@@ -3,6 +3,9 @@ package com.tavemakers.surf.domain.reservation.repository;
 import com.tavemakers.surf.domain.reservation.entity.Reservation;
 import com.tavemakers.surf.domain.reservation.entity.ReservationStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -17,6 +20,8 @@ public interface ReservationRepository extends JpaRepository<Reservation, Long> 
 
     Optional<Reservation> findByPostIdAndStatus(Long postId, ReservationStatus status);
 
-    void deleteByPostId(Long postId);
+    @Modifying(clearAutomatically = true, flushAutomatically = true)
+    @Query("DELETE FROM Reservation r WHERE r.postId = :postId")
+    void deleteByPostId(@Param("postId") Long postId);
 
 }
